Cleanup positioning and usage
